requirements.txt
###### Requirements to be installed ######
## pip install -r requirements.txt
requests
bs4
sentence_transformers
faiss-cpu
numpy
groq
